got particles to work, now styling of main page
