+++ client/package.json
@@ -7,20 +7,21 @@
     "@apollo/react-hooks": "^4.0.0",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
-    "apollo-server-express": "^3.5.0",
-    "react": "^17.0.2",
     "apollo-boost": "^0.4.9",
+    "apollo-server-express": "^3.5.0",
     "bootstrap": "^5.1.3",
     "graphql": "^16.0.1",
     "graphql-tag": "^2.12.6",
     "jwt-decode": "^3.1.2",
+    "react": "^17.0.2",
     "react-apollo": "^3.1.5",
     "react-bootstrap": "^2.0.2",
     "react-dom": "^17.0.1",
+    "react-native-linear-gradient": "^2.5.6",
     "react-particles-js": "^3.6.0",
     "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.2",
-    "react-tsparticles": "^1.37.3"
+    "react-tsparticles": "^1.37.4"
   },
   "scripts": {
     "start": "react-scripts start",
