+++ client/src/App.css
@@ -1,5 +1,3 @@
-/* will input background gradient here when we go over it tomorrow  */
-
 .App {
   text-align: center;
 }
@@ -28,4 +26,13 @@
 
 .App-link {
   color: #61dafb;
+}
+
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
 }
