--- client/src/ParticleBackground.js
@@ -1,9 +1,0 @@
-import React from "react";
-import Particles from "react-particles-js";
-import ParticleConfig from "../src/config/particle-config";
-
-export default function ParticleBackground() {
-    return (
-        <Particles params={ParticleConfig}></Particles>
-    );
-}
