--- client/src/config/particle-config.js
@@ -1,112 +1,0 @@
-const particlesConfig = {
-    particles: {
-      number: {
-        value: 400,
-        density: {
-          enable: false,
-          value_area: 1000
-        }
-      },
-      color: {
-        value: "#7d6587"
-      },
-      shape: {
-        type: "circle",
-        stroke: {
-          width: 0,
-          color: "#000000"
-        },
-        polygon: {
-          nb_sides: 5
-        },
-        image: {
-          src: "img/github.svg",
-          width: 100,
-          height: 100
-        }
-      },
-      opacity: {
-        value: 1,
-        random: false,
-        anim: {
-          enable: true,
-          speed: 0.1,
-          opacity_min: 0,
-          sync: false
-        }
-      },
-      size: {
-        value: 2,
-        random: true,
-        anim: {
-          enable: false,
-          speed: 4,
-          size_min: 0.3,
-          sync: false
-        }
-      },
-      line_linked: {
-        enable: false,
-        distance: 16.03412060865523,
-        color: "#aaa1a1",
-        opacity: 0.4,
-        width: 1
-      },
-      move: {
-        enable: true,
-        speed: 1,
-        direction: "bottom-right",
-        random: true,
-        straight: true,
-        out_mode: "out",
-        bounce: false,
-        attract: {
-          enable: true,
-          rotateX: 600,
-          rotateY: 600
-        }
-      }
-    },
-    interactivity: {
-      detect_on: "canvas",
-      events: {
-        onhover: {
-          enable: true,
-          mode: "repulse"
-        },
-        onclick: {
-          enable: true,
-          mode: "repulse"
-        },
-        resize: true
-      },
-      modes: {
-        grab: {
-          distance: 400,
-          line_linked: {
-            opacity: 1
-          }
-        },
-        bubble: {
-          distance: 250,
-          size: 0,
-          duration: 2,
-          opacity: 0,
-          speed: 3
-        },
-        repulse: {
-          distance: 100,
-          duration: 0.4
-        },
-        push: {
-          particles_nb: 4
-        },
-        remove: {
-          particles_nb: 2
-        }
-      }
-    },
-    retina_detect: true
-  };
-
-export default particlesConfig;
